删除无关代码

+++ src/main/java/com/ogc/standard/bo/IHLOrderBO.java
@@ -1,6 +1,5 @@
 package com.ogc.standard.bo;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 import com.ogc.standard.bo.base.IPaginableBO;
@@ -11,7 +10,7 @@ import com.ogc.standard.enums.EHLOrderStatus;
 
 public interface IHLOrderBO extends IPaginableBO<HLOrder> {
 
-    String applyOrder(Account account, Jour jour, BigDecimal applyAmount,
+    String applyOrder(Account account, Jour jour, Long applyAmount,
             String applyUser, String applyNote);
 
     void approveOrder(HLOrder order, EHLOrderStatus status, String approveUser,
